requirements.txt
PyQt5==5.15.10
sentence-transformers==2.2.2
chromadb==0.4.22
gpt4all==2.0.2
pdfplumber==0.10.3
python-docx==1.1.0
torch==2.1.2
